refactor: adjusment router
