+++ src/components/Home/CategoriesSection.jsx
