+++ src/components/Partials/Layout.jsx
@@ -1,8 +1,7 @@
 import { useState } from "react";
-import DiscountBanner from "../Home/DiscountBanner";
 import Drawer from "../Mobile/Drawer";
 import Footer from "./Footers/Footer";
-import Header from "./Headers/HeaderOne";
+import Header from "./Headers/Header";
 
 export default function Layout({ children, childrenClasses }) {
   const [drawer, setDrawer] = useState(false);
@@ -14,7 +13,7 @@ export default function Layout({ children, childrenClasses }) {
         <div className={`w-full  ${childrenClasses || "pt-[30px] pb-[60px]"}`}>
           {children && children}
         </div>
-        <DiscountBanner />
+        {/* <DiscountBanner /> */}
         <Footer />
       </div>
     </>
